% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.lpp.R
\name{angle.lpp}
\alias{angle.lpp}
\title{Orientation angles of lines connecting points over a linear network}
\usage{
angle.lpp(X, degree = TRUE, directed = TRUE)
}
\arguments{
\item{X}{an object of class lpp}

\item{degree}{logical. if TRUE, output is based on angles from 0 to 360}

\item{directed}{logical. if TRUE, the algorithm respects the sense of direction, see \link[spatstat]{angles.psp}}
}
\value{
a matrix of angles
}
\description{
Calculates the orientation angle of lines connecting points over a linear network
}
\details{
This function calculates the pairwise orientation between points over a linear network.
}
\examples{
# generate random relaisations
X <- rpoislpp(0.001,branchnet)
M <- angle.lpp(X)
}
\references{
Moradi, M., Mateu, J,. and Comas, C. (2020) Directional analysis for point patterns on linear networks. Stat.
}
\seealso{
\link[spatstat]{angles.psp}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
