% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisLinearKinhom.R
\name{anisLinearKinhom}
\alias{anisLinearKinhom}
\title{Inhomogeneous K-function accounting for directions}
\usage{
anisLinearKinhom(
  X,
  lambda = lambda,
  normalize = TRUE,
  r = NULL,
  phi = NULL,
  maxphi = 360,
  edge = TRUE,
  nxy = 10,
  parallel = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{an object of class lpp}

\item{lambda}{estimated intensity at data points. If not given, it will be automatically obtained from \link[spatstat]{densityQuick.lpp}}

\item{normalize}{logical. whether to normalize the estimate. See \link[spatstat]{linearKinhom}}

\item{r}{optional. distance vector wherein the K-funtion is evaluated}

\item{phi}{optional. angle vector wherein the K-funtion is evaluated}

\item{maxphi}{max angle. this will be used only when phi is not given}

\item{edge}{optional. whether to use edge correction or not}

\item{nxy}{dimension. this will be considered as the length of r and phi when they are not given}

\item{parallel}{logical. if TRUE, it uses \link[parallel]{mclapply}.}

\item{verbose}{logical. if TRUE, it prints the progress of the function.}

\item{...}{arguments passed to \link[parallel]{mclapply} for configuration of the cores.}
}
\value{
an object of class sumlpp
}
\description{
Inhomogeneous K-function accounting for directions
}
\details{
This function calculates the inhomogeneous K-function over a grid of distances and angles.
}
\note{
this function can be quite slow, we suggest to use \link[anisotropylpp]{nnangle.lpp}
}
\examples{
# generate random relaisations
X <- rpoislpp(0.001,branchnet)
K <- anisLinearKinhom(X)
}
\references{
Moradi, M., Mateu, J,. and Comas, C. (2020) Directional analysis for point patterns on linear networks. Stat.
}
\seealso{
\link[spatstat]{linearKinhom}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
