% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immig.death.lpp.R
\name{Immig.death.lpp}
\alias{Immig.death.lpp}
\title{Simulate immigration-death pattern over  a linear network}
\usage{
Immig.death.lpp(
  L,
  a = pi/2,
  t = 10,
  alpha = 1,
  d = 0.001,
  prob_accep = 1,
  r = 0.2,
  N = 0,
  strauss = FALSE,
  burnin = 10000,
  nsim = 1,
  verbose = TRUE,
  Figure = TRUE,
  C = 1
)
}
\arguments{
\item{L}{a linear network}

\item{a}{anisotropy angle}

\item{t}{pi/t is the tolerance}

\item{alpha}{immigration arte}

\item{d}{death rate}

\item{prob_accep}{probability to accept a new arriving point}

\item{r}{r}

\item{N}{N}

\item{strauss}{strauss}

\item{burnin}{burnin}

\item{nsim}{number of patterns to be simulated}

\item{verbose}{logical}

\item{Figure}{logical}

\item{C}{C}
}
\value{
if \code{nsim=1} a single realization, otherwise a list of realizations of lpp objects
}
\description{
Simulate immigration-death pattern over  a linear network
}
\examples{
 X <- Immig.death.lpp(L=branchnet,a=pi/2,t=10,alpha=1.0,d=0.001,prob_accep=0,r=0.15,N=0,strauss=FALSE,burnin=10000,nsim=1,verbose=TRUE,Figure=TRUE,C=1)
 plot(X,pch=20,col="red",cols="blue",main="")
}
\references{
Moradi, M., Mateu, J,. and Comas, C. (2020) Directional analysis for point patterns on linear networks. Stat.
}
\author{
Carles Comas \email{carles.comas@udl.cat}, Mehdi Moradi \email{m2.moradi@yahoo.com}
}
