% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnangle.lpp.R
\name{nnangle.lpp}
\alias{nnangle.lpp}
\title{Nearest neighbor orientation}
\usage{
nnangle.lpp(X, k = 1, degree = TRUE, directed = TRUE)
}
\arguments{
\item{X}{an object of class lpp}

\item{k}{The algorithm will find the angle towards the kth nearest neighbor}

\item{degree}{logical. if TRUE, output is based on angles}

\item{directed}{logical. if TRUE, the algorithm considers direction}
}
\value{
a vector of angles
}
\description{
Calculates the angle to the nearest neighbor point over a linear network
}
\details{
this function calculates the orientation to the kth nearest neighbor for each point
}
\examples{
# generate random relaisations
X <- rpoislpp(0.001,branchnet)
nnange <- nnangle.lpp(X)
}
\references{
Moradi, M., Mateu, J,. and Comas, C. (2020) Directional analysis for point patterns on linear networks. Stat.
}
\seealso{
\link[spatsat]{nndist}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
