% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.anis.MatClust.lpp.R
\name{r.anis.MatClust.lpp}
\alias{r.anis.MatClust.lpp}
\title{Simulate anisotropic Matern cluster process on linear networks}
\usage{
r.anis.MatClust.lpp(
  pint = 10,
  r = 0.2,
  cint = 10,
  L = L,
  a = pi/3,
  t = 10,
  model = c("Poisson", "Uniform"),
  relaxed = FALSE,
  prob = 0.5,
  nsim = 1,
  check = FALSE
)
}
\arguments{
\item{pint}{intensity function of the parents if model is Poisson, otherwise the number of parrents}

\item{r}{iteraction distance}

\item{cint}{intensity function of offsprings if model is Poisson, otherwise the number of offsprings per cluster}

\item{L}{a linear network}

\item{a}{angle of interaction}

\item{t}{pi/t will be considered as a tolerance for angle of interaction}

\item{model}{model to be considered for simulating parents and offsprings, currently only Poisson and Uniform. Default is Poisson.}

\item{relaxed}{logical. if TRUE the function keeps some points which do not fall within angle condition according to a given probability}

\item{prob}{if relaxed=TRUE, then the simulation also includes points, which do not fall within an angle a with respect to their parent, with probability prob}

\item{nsim}{number of simulations}

\item{check}{logical}
}
\value{
if \code{nsim=1} a single realization, otherwise a list of realizations of lpp objects
}
\description{
Simulate anisotropic Matern cluster process on linear networks based on given interaction distance and angle
}
\details{
this function generates realizations from (an)isotropic Matern Cluster point processes over linear networks. the strength of anisotropy can be controlled by \code{relaxed} and \code{prob}
}
\examples{
# generate random realizations
X <- r.anis.MatClust.lpp(pint=0.01,r=300,cint=0.2,L=branchnet,t=50)
plot(X,main="",cols=2,col=4,pch=20)
Y <- r.anis.MatClust.lpp(pint=0.01,r=300,cint=0.2,L=branchnet,t=50,a=pi/2)
plot(Y,main="",cols=2,col=4,pch=20)
}
\references{
Moradi, M., Mateu, J,. and Comas, C. (2020) Directional analysis for point patterns on linear networks. Stat.
}
\seealso{
\link[spatstat]{rMatClust}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
